+++ KOR_s12181_2/src/edu/pjwstk/kor/model/data/RandomPaymentType.java
@@ -36,29 +36,29 @@ public class RandomPaymentType {
 		paymentTypeNames.add("Koperta du¿a");
 		paymentTypeNames.add("Koperta max");
 		
-		List<Float> paymentPrices = new ArrayList<Float>();
-		paymentPrices.add(Float.parseFloat("12.99"));
-		paymentPrices.add(Float.parseFloat("16.99"));
-		paymentPrices.add(Float.parseFloat("25.99"));
-		paymentPrices.add(Float.parseFloat("69.99"));
-		paymentPrices.add(Float.parseFloat("10"));
-		paymentPrices.add(Float.parseFloat("25.99"));
-		paymentPrices.add(Float.parseFloat("33.99"));
-		paymentPrices.add(Float.parseFloat("51.99"));
-		paymentPrices.add(Float.parseFloat("129.99"));
-		paymentPrices.add(Float.parseFloat("25"));
-		paymentPrices.add(Float.parseFloat("5.99"));
-		paymentPrices.add(Float.parseFloat("7.99"));
-		paymentPrices.add(Float.parseFloat("11.99"));
+		List<Double> paymentPrices = new ArrayList<Double>();
+		paymentPrices.add(Double.parseDouble("12.99"));
+		paymentPrices.add(Double.parseDouble("16.99"));
+		paymentPrices.add(Double.parseDouble("25.99"));
+		paymentPrices.add(Double.parseDouble("69.99"));
+		paymentPrices.add(Double.parseDouble("10"));
+		paymentPrices.add(Double.parseDouble("25.99"));
+		paymentPrices.add(Double.parseDouble("33.99"));
+		paymentPrices.add(Double.parseDouble("51.99"));
+		paymentPrices.add(Double.parseDouble("129.99"));
+		paymentPrices.add(Double.parseDouble("25"));
+		paymentPrices.add(Double.parseDouble("5.99"));
+		paymentPrices.add(Double.parseDouble("7.99"));
+		paymentPrices.add(Double.parseDouble("11.99"));
 		
-		paymentPrices.add(Float.parseFloat("8.99"));
-		paymentPrices.add(Float.parseFloat("12.99"));
-		paymentPrices.add(Float.parseFloat("19.99"));
-		paymentPrices.add(Float.parseFloat("0.0"));		
-		paymentPrices.add(Float.parseFloat("2.99"));
-		paymentPrices.add(Float.parseFloat("3.99"));
-		paymentPrices.add(Float.parseFloat("4.99"));
-		paymentPrices.add(Float.parseFloat("0.0"));
+		paymentPrices.add(Double.parseDouble("8.99"));
+		paymentPrices.add(Double.parseDouble("12.99"));
+		paymentPrices.add(Double.parseDouble("19.99"));
+		paymentPrices.add(Double.parseDouble("0.0"));		
+		paymentPrices.add(Double.parseDouble("2.99"));
+		paymentPrices.add(Double.parseDouble("3.99"));
+		paymentPrices.add(Double.parseDouble("4.99"));
+		paymentPrices.add(Double.parseDouble("0.0"));
 		
 		for (int i=0; i<21; i++) {
